+++ jctool/com-builder/com_builder.hpp
@@ -1,0 +1,2 @@
+#pragma once
+#include "../con_com.hpp"
