--- jctool/commands/com_builder.hpp
@@ -1,2 +1,0 @@
-#pragma once
-#include "con_com.hpp"
