+++ jctool/include/jctool_leds.hpp
@@ -1,0 +1,6 @@
+#pragma once
+#include "jctool_types.h"
+
+namespace LEDS {
+    int set_led_busy(CT& ct, ConHID::ProdID con_type);
+}
