+++ jctool/include/jctool_nfc.hpp
@@ -1,0 +1,6 @@
+#pragma once
+#include "jctool_types.h"
+
+namespace NFC {
+    int nfc_tag_info(CT& ct);
+}
